feat(UIPointer): automatically activate UI ray within given distance
A new option has been added to the UI Pointer that will automatically
activate the UI raycast when the game object containing a UI Pointer
gets within a given distance of a valid canvas.

The PlayerObject script has been updated as well to be able to
identify player objects that are considered colliders.